use grupo-only tags to tag grupos in Grupo seeder
Grupos should only be tagged by taggable tags. Some tags are used to tag
grupos, while others are used to tag other tags. For example,  the  tags
POLI  and  EACH  shall  be  used  to  tag  grupos  belonging  to   these
departaments;  the  tags  POLI  and  EACH  are   tagged   by   the   tag
"departamento"; it does not make  sense  to  tag  a  grupo  by  the  tag
"departamento", since it is a tag-only tag.